+++ tests/testthat/test-helper-fake_lbk.R
@@ -17,6 +17,6 @@ test_that("fake_lbk creates new columns", {
   )
 })
 
-test_that("fake_lbk outputs known value", {
-  expect_known_value(fake_lbk(), "ref-helper-fake_lbk", update = FALSE)
+test_that("fake_lbk outputs expected value", {
+  expect_snapshot_value(fake_lbk(), style = "json2")
 })