+++ tests/testthat/test-restructure_ald.R
@@ -46,6 +46,6 @@ test_that("correctly overwrites sector", {
 
 test_that("restructure_ald outputs the expected tibble", {
   out <- restructure_ald(fake_ald())
-  expect_is(out, "tbl_df")
+  expect_s3_class(out, "tbl_df")
   expect_named(out, c("name", "sector", "alias"))
 })