+++ NEWS.md
@@ -4,6 +4,7 @@
 
 # r2dii.match 0.0.6
 
+* Use snapshots via testthat 3e.
 * `match_names()` gains `...` to pass additional arguments to
   `stringdist::stringsim()` (@evgeniadimi #310).
 * `prioritize()` with 0-row input now returns the input untouched (#284).